Searching for packages on CRAN

Searches for packages on CRAN by scanning a specified set of information fields for a user-provided search term.

exploreFields(
  term,
  fields = c("Name", "Description", "LongDescription"),
  mode = "or",
  match = "like",
  display = "viewer",
  index = NULL
)

Arguments

term
Search term to look for; character vector must have one element.

fields
The list of fields to be scanned for the search term; must be a character vector with one or more field names. Allowed field names are: "Name", "Description", "LongDescription", "Maintainer", "Authors@R", "Author", "License", "Imports", "Enhances", "Depends", "Suggests", "Reverse depends", "Reverse suggests", "Reverse enhances", "Copyright", "Contact", "Note", "MailingList".

mode
Indicates whether matches in the field shall be combined with a logical OR or with a logical AND; accordingly, permitted values are "or" (default) and "and". In "or" mode, every package that has the search term in any of the fields from fields generates a hit, in "and" mode the search term must be found in all fields provided to make that package a search hit.

match
Either "like" (default) or "exact". Determines if the field content must match the search term exactly or only needs to contain it.

display
Describes where the search results shall be shown. Either "viewer" or "console". If "viewer", the results are shown in RStudio's Viewer pane if the RStudio IDE is being used). If results = "console" the search results are shown as a text table in the R console.

index
Either a path (or URL) to a search index, or a search index that is already loaded. If no index is provided, exploreFields() creates an ad hoc search index.

Value

No return value.

Details

Found packages are listed in alphabetical order, there is no prioritization of search hits as in findPackage().

Author

Joachim Zuckarelli joachim@zuckarelli.de

Examples

 exploreFields("Hadley", c("Maintainer", "Authors@R", "Author")) 

#> 
#> Results: 152 out of 17409 CRAN packages found in 5 seconds.