Showing information about a package

Shows detailed CRAN information for a package.

packageDetails(package, brief = FALSE, show.tip = TRUE, index = NULL)

Arguments

package
Either the name of a package (capitalization does generally not matter) or the search result number shown in the results of findPackage() (the number in the GO column). Only one package is allowed here.

brief
If "TRUE", only title, short and long description as well as the maintainer of the package are shown, otherwise all available fields are displayed.

show.tip
If "TRUE", tips for getting additional information on the package are shown.

index
Either a path (or URL) to a search index, or a search index that is already loaded. If no index is provided, packageDetails() creates an ad hoc search index.

Value

No return value.

Author

Joachim Zuckarelli joachim@zuckarelli.de

Examples

 packageDetails("ggplot2") 

#> 
#> Package ggplot2 
#> 
#> Title: Create Elegant Data Visualisations Using the Grammar of Graphics
#> 
#> Long description: 
#> A system for 'declaratively' creating graphics,
#>     based on "The Grammar of Graphics". You provide the data, tell 'ggplot2'
#>     how to map variables to aesthetics, what graphical primitives to use,
#>     and it takes care of the details.
#> 
#> Publish date: 2020-12-30
#> 
#> Version: 3.3.3
#> 
#> License: MIT + file LICENSE
#> 
#> Maintainer: Thomas Lin Pedersen <thomas.pedersen@rstudio.com>
#> 
#> URL: https://ggplot2.tidyverse.org,
#> https://github.com/tidyverse/ggplot2
#> 
#> Bug reports: https://github.com/tidyverse/ggplot2/issues
#> 
#> Authors: 
#> Hadley Wickham [aut] (<https://orcid.org/0000-0003-4757-117X>),
#>   Winston Chang [aut] (<https://orcid.org/0000-0002-1576-2126>),
#>   Lionel Henry [aut],
#>   Thomas Lin Pedersen [aut, cre]
#>     (<https://orcid.org/0000-0002-5147-4711>),
#>   Kohske Takahashi [aut],
#>   Claus Wilke [aut] (<https://orcid.org/0000-0002-7470-9261>),
#>   Kara Woo [aut] (<https://orcid.org/0000-0002-5125-4188>),
#>   Hiroaki Yutani [aut] (<https://orcid.org/0000-0002-3385-7233>),
#>   Dewey Dunnington [aut] (<https://orcid.org/0000-0002-9415-4582>),
#>   RStudio [cph, fnd]
#> 
#> Authors@R: 
#> c(
#>     person("Hadley", "Wickham", , "hadley@rstudio.com", "aut",
#>            comment = c(ORCID = "0000-0003-4757-117X")),
#>     person("Winston", "Chang", , role = "aut",
#>            comment = c(ORCID = "0000-0002-1576-2126")),
#>     person("Lionel", "Henry", , role = "aut"),
#>     person("Thomas Lin", "Pedersen", , "thomas.pedersen@rstudio.com", 
#>            role = c("aut", "cre"), comment = c(ORCID = "0000-0002-5147-4711")),
#>     person("Kohske", "Takahashi", role = "aut"),
#>     person("Claus", "Wilke", role = "aut",
#>            comment = c(ORCID = "0000-0002-7470-9261")),
#>     person("Kara", "Woo", role = "aut",
#>            comment = c(ORCID = "0000-0002-5125-4188")),
#>     person("Hiroaki", "Yutani", role = "aut",
#>            comment = c(ORCID = "0000-0002-3385-7233")),
#>     person("Dewey", "Dunnington", role = "aut",
#>            comment = c(ORCID = "0000-0002-9415-4582")),
#>     person("RStudio", role = c("cph", "fnd"))
#>     )
#> 
#> Enhances: sp
#> 
#> Imports: 
#> digest, glue, grDevices, grid, gtable (>= 0.1.1), isoband,
#> MASS, mgcv, rlang (>= 0.3.0), scales (>= 0.5.0), stats, tibble,
#> withr (>= 2.0.0)
#> 
#> Depends: R (>= 3.2)
#> 
#> Suggests: 
#> covr, dplyr, ggplot2movies, hexbin, Hmisc, knitr, lattice,
#> mapproj, maps, maptools, multcomp, munsell, nlme, profvis,
#> quantreg, RColorBrewer, rgeos, rmarkdown, rpart, sf (>= 0.7-3),
#> svglite (>= 1.2.0.9001), testthat (>= 2.1.0), vdiffr (>= 0.3.0)
#> 
#> Reverse depends: 
#> ACSNMineR, AFheritability, afmToolkit, alakazam, alookr, AmpliconDuo, Anaquin, AneuFinder, apisensr, applicable, bacon, BasketballAnalyzeR, bayesDP, BCellMA, bde, BDMMAcorrect, bhm, biclustermd, bmscstan, bootnet, bpcp, braidReports, btergm, bunching, caret, CellNOptR, centralplot, ceterisParibus, changepoint.geo, ChaosGame, CHETAH, ChIPQC, chromstaR, cjoint, classifierplots, climwin, clustEff, clustrd, CNVrd2, CNVScope, coefplot, cogena, cohorttools, corkscrew, CoSMoS, CountClust, CrispRVariants, crmPack, Crossover, CRTgeeDR, cummeRbund, curatedBreastData, cystiSim, cytofan, dae, DaMiRseq, dampack, dartR, dbmss, decompTumor2Sig, DeconRNASeq, Deducer, deltaGseg, DendroSync, DengueRT, DepthProc, DEqMS, dfexplore, DHBins, diathor, diffEnrich, diffeR, dittoSeq, DoTC, dotwhisker, dslice, dtwSat, dvqcc, dynr, Eagle, earlywarnings, ecap, eeptools, egg, EnhancedVolcano, EnsCat, entropart, EpiCurve, epihet, episensr, erccdashboard, EventStudy, FactoClass, factoextra, Factoshiny, fbroc, FisherEM, fishmove, flippant, forestmodel, FourCSeq, freqparcoord, frequency, func2vis, funcharts, FunciSNP, funLBM, gapfill, garma, GARS, gcerisk, gde, GENEAsphere, genlogis, genomeplot, GenomicOZone, geomnet, geotoolsR, gespeR, ggallin, ggalluvial, GGally, ggalt, gganimate, ggbeeswarm, ggbio, ggbuildr, ggcharts, ggcorrplot, ggcyto, ggdag, ggDCA, ggdemetra, ggedit, ggetho, ggfocus, ggforce, ggformula, ggfortify, ggghost, gggibbous, ggh4x, gghalves, gghighlight, ggimage, gginnards, ggip, gglorenz, ggmap, ggmcmc, ggmosaic, ggmulti, ggnetwork, ggOceanMaps, ggparallel, ggparty, ggpmisc, ggpol, ggpolypath, ggpubr, ggraph, ggraptR, ggrepel, ggROC, ggseas, ggside, ggsn, ggspectra, ggswissmaps, ggtern, ggTimeSeries, ggvenn, ggvoronoi, ggwordcloud, glow, GOplot, GOxploreR, gpmap, granovaGG, GRENITS, GSCA, gsDesign, gstar, Guitar, GWAS.BAYES, gwdegree, h5vc, HCmodelSets, HiLDA, Hmisc, HRM, hyperSpec, iarm, iBMQ, iCARH, iCellR, ICSsmoothing, ICtest, idm, ifaTools, IMAS, infinitefactor, inlabru, InSilicoVA, interplot, inti, invacost, ioanalysis, IPDfromKM, IsoformSwitchAnalyzeR, itan, itemanalysis, iteremoval, itraxR, IVAS, kayadata, LambertW, lavaSearch2, lcc, learnstats, likeLTD, likert, LinkHD, listdtr, llbayesireg, logihist, longRPart2, loon.ggplot, lsbclust, MAFDash, malariaAtlas, maser, MCAvariants, mccf1, MCMC.OTU, MCMC.qpcr, mcprofile, metacart, metaforest, meteogRam, methimpute, methylumi, mgcViz, mhtboot, microbiome, MicroNiche, MIMOSA, MinEDfind, MineICA, missRows, MIXFIM, mixKernel, mixOmics, MixOptim, MLSeq, mlxR, mnreadR, monocle, MortalityTables, motoRneuron, mrfDepth, multilevelPSA, MultisiteMediation, NanoMethViz, NBLDA, Nebulosa, NetworkChange, ngsReports, NIRStat, nonmem2R, npde, oceanwaves, OptimalRerandExpDesigns, optimos.prime, orgR, OriGen, overlapping, pafr, pagoo, PairedData, pAnalysis, PASenseWear, PAsso, PASWR2, pathVar, pauwels2014, PCAtools, PedCNV, pequod, perry, personalized, perspectev, phase1PRMD, phase1RMD, phenModel, phenoTest, photosynthesis, phyloseqGraphTest, Pi, pitchRx, pkggraph, PKreport, plantecowrap, PlasmaMutationDetector, plothelper, plotly, plotROC, pointRes, PortfolioEffectHFT, PPInfer, PPtreeViz, precintcon, primer, PRISMA, ProbBayes, profileR, PupillometryR, pvldcurve, qqplotr, qrqc, quadrupen, QualInt, quickpsy, rabhit, radiant.data, RAM, randomizeR, RATest, RatingScaleReduction, RcmdrPlugin.KMggplot2, rdacca.hp, RDAVIDWebService, recoup, reinsureR, Repliscope, rise, rms, RNAseqNet, RNASeqR, RnBeads, Rnits, robCompositions, robmed, robustHD, rorutadis, RpeakChrom, Rqc, rrecsys, RSA, rSEA, RsSimulx, rstan, rtip, rwty, samExploreR, savR, scAlign, scater, scatterpie, schex, scoper, SCOUTer, scPipe, scRepertoire, SDLfilter, sdStaf, season, SEERaBomb, selfea, sensiPhy, seqsetvis, SetMethods, sglr, shazam, shinySIR, simmer.plot, simmr, SmarterPoland, SMFI5, soc.ca, SparseBiplots, SparseFactorAnalysis, sparsereg, spatialClust, SPEI, spqn, sprm, ssviz, staRdom, StatMatch, statVisual, sugrrants, SummarizedBenchmark, survHE, survminer, survRatio, swa, SWMPr, synlet, TarSeqQC, TcGSA, tdr, texmex, texteffect, tigger, timeline, TissueEnrich, TPD, TreeDep, TriMatch, TripleR, tspmeta, twangRDC, useful, ushr, varian, variancePartition, vdg, visae, VplotR, vrcp, waffle, weightr, WRTDStidal, xkcd, Xplortext, xpose, YAPSA, zooaRch
#> 
#> Reverse suggests: 
#> abctools, abd, activatr, activityCounts, AdaptGauss, adept, adklakedata, adoptr, afex, aggregateBioVar, aghq, agridat, ahpsurvey, aire.zmvm, airGR, airt, ALA4R, alfred, AlleleShift, AmpGram, AMR, amt, antaresViz, antitrust, aopdata, apsimx, ar.matrix, archetypes, archivist, arcos, arctools, ARPobservation, ARTool, ashr, asymptor, AugmenterR, automap, automultinomial, autothresholdr, av, avocado, babette, bacondecomp, bambu, bang, baRulho, BAS, base2grob, BaseSet, BatchGetSymbols, bayesbio, bayestestR, bbmle, bcdata, BCEA, bcmaps, bcp, beautier, belg, bench, benchmarkfdrData2019, benchmarkme, benchr, benthos, bestNormalize, BetaBit, bfw, bib2df, BiDimRegression, binom, binomialMix, binsegRcpp, bioassayR, biobroom, biogas, biogram, Bioi, bioset, Biostatistics, BioTIP, BIS, bisect, bitmexr, BlandAltmanLeh, blastula, blockCV, blockRAR, BMRSr, bodenmiller, BOJ, bomrang, bootUR, borrowr, brainGraph, breathteststan, BRGenomics, bridgedist, briskaR, brokenstick, broom, broom.helpers, broom.mixed, brotli, brunnermunzel, BSDA, bsem, bssm, bsub, bulletcp, bwsTools, cancensus, CancerGram, cansim, caRecall, CarletonStats, cate, CDatanet, CellaRepertorium, CellBench, cellHTS2, celltrackR, Census2016, CGGP, cheatR, checkmate, cherryblossom, childsds, chilemapas, ChillModels, choroplethrMaps, chorrrds, chron, ciftiTools, ciTools, civis, ClassifyR, CLLmethylation, clusternomics, ClusterSignificance, cmocean, codingMatrices, codyn, collapse, colorednoise, colorspace, comparer, COMPASS, compboost, comtradr, consensusOV, consensusSeekeR, constellation, contactdata, contextual, contoureR, ContourFunctions, contrast, CoRegFlux, CornerstoneR, correlation, countytimezones, covafillr, CPAT, cpp11, cpsvote, crandep, crawl, CropDetectR, crosstable, crosstalk, cSEM, CTD, curatedAdipoChIP, curatedAdipoRNA, curatedMetagenomicData, cusum, CVEK, cvequality, cyclomort, Cyclops, czechrates, DAISIE, dams, danstat, dasper, dataresqc, dataRetrieval, datarobot, datasauRus, datplot, daymetr, DeclareDesign, deconvolveR, DeepBlueR, deGradInfer, densvis, denvax, depmap, derfinder, derivmkts, descstat, designr, detrendr, dexter, dexterMST, DGCA, DGVM3D, dielectric, diffuStats, DImodels, dimRed, dint, DirectEffects, directlabels, disclapmix, discrim, diseq, DisImpact, disk.frame, distributions3, dobin, dodgr, dominanceanalysis, dpseg, drake, drawer, DrBats, dscore, DSjobtracker, dstack, dynutils, earthtones, EbayesThresh, ec50estimator, Ecfun, echo.find, echor, ecoCopula, ecotox, edfReader, edgarWebR, edge, eemR, eesim, EFDR, effectsize, EGAnet, EHR, eia, embed, emmeans, EMMIXmfa, emuR, enrichwith, eph, epigraphdb, epiR, EpiTxDb, eq5d, equatiomatic, escalation, esetVis, etl, etm, eurlex, europepmc, eurostat, evaluate, exact2x2, expss, ExtDist, eyelinker, fable.prophet, facerec, factory, fastRG, FDboost, fddm, fec16, fedregs, feisr, fermicatsR, FField, fgeo.analyze, fhidata, fitbitScraper, fitdistrplus, fitHeavyTail, fitzRoy, fivethirtyeight, flacco, flexrsurv, flexsurv, flextable, flowTime, FMM, forcats, forrel, forwards, foster, fourierin, fractional, fredr, frequencyConnectedness, FRESA.CAD, frontiles, frost, FSelectorRcpp, functiondepends, funData, funrar, funtimes, furrowSeg, fuser, FusionLearn, fuzzyjoin, gaiah, ganalytics, GAPGOM, gapminder, gcatest, GDAdata, gear, gender, geneplast, GENESIS, GenoGAM, GenomicDataCommons, genphen, geobr, geodiv, geofi, geoknife, GetBCBData, getCRUCLdata, GetDFPData, GetDFPData2, GetLattesData, GetQuandlData, GetTDData, ggeffects, ggROC, ghcm, ghibli, ghypernet, gifski, GillespieSSA2, git2rdata, glm.predict, GLMcat, glmGamPoi, glmmTMB, glmpca, glue, googlesheets, gplite, grainchanger, graphhopper, graphlayouts, grates, grattan, Greg, gridDebug, gridExtra, GROAN, groupdata2, groupedstats, growthcurver, GSgalgoR, gsloid, GSODR, GSVA, gtable, guardianapi, gutenbergr, h2o, hal9001, Haplin, harrietr, hazus, hbbr, HDCytoData, hdfqlr, HDiR, heddlr, hereR, hermiter, heuristica, HIBAG, highcharter, hilbertSimilarity, hIRT, HistData, historydata, hmclearn, HMP16SData, HMP2Data, holodeck, horseshoe, hpa, hsrecombi, HTLR, httk, hues, hutils, huxtable, hydroscoper, hystReet, i2dash, ialiquor, icarus, ICON, iconr, icr, icrf, idiogramFISH, IgGeneUsage, igoR, IHW, imagefluency, imager, imgrec, imputeFin, installr, InterVA5, iNZightPlots, ipmisc, ipumsr, irg, isoband, isocat, ivmte, JacobiEigen, jmvcore, JointAI, kamila, keras, kergp, kfigr, kibior, kiwisR, klsh, knitrBootstrap, KSD, LAGOSNE, Lahman, latex2exp, lda, LDheatmap, ldsr, LEAP, lfa, lgrdata, lillies, lingtypology, listdown, lme4, LocalControl, localIV, logitnorm, logitr, lognorm, longCatEDA, loo, LOPART, lotri, LPWC, LSMRealOptions, lspline, ltbayes, lutz, MAc, macleish, MACPET, MAd, maddison, madrat, magick, MANOVA.RM, mapchina, mapi, mapsapi, mapsFinland, margins, marked, marketr, marqLevAlg, MARSS, matchingMarkets, matrixStats, matrixStrucTest, matsindf, MBNMAdose, MBNMAtime, mcbette, medicalrisk, medicare, memor, metagenomeFeatures, MetaStan, methcon5, methylCC, metricsgraphics, mets, mfGARCH, mgc, MGDrivE2, MGLM, microbenchmark, mifa, miic, mikropml, milr, mindicador, missMethods, mistat, mistr, mixdir, mixIndependR, MixMatrix, mleap, mlogit, mlr3pipelines, mlr3proba, mlrMBO, MLVSBM, mmb, mmpf, MNLpred, modcmfitr, modelbased, modelr, modelsummary, MODISTools, MODIStsp, ModStatR, MOEADr, moexer, momentuHMM, MoMPCA, mongolite, MonteCarlo, moonBook, mortAAR, mosaic.find, mosaicData, mosmafs, MOSS, move, mrbayes, MrSGUIDE, MSG, mstate, mthapower, MTLR, MTPS, MultiBD, multiROC, munsell, musica, mvSLOUCH, mwaved, mwshiny, nanny, nasapower, naspaclust, nationwider, NBAloveR, NBAMSeq, nbapalettes, nbTransmission, nc, NestLink, NetMix, netrankr, netSmooth, networkABC, neurobase, NewWave, nhdplusTools, NHSRdatasets, NitrogenUptake2016, nlraa, nls.multstart, NlsyLinks, nLTT, NNTbiomarker, nntrf, noaaoceans, nomisr, nosoi, NPARC, nprcgenekeepr, npsurvSS, nseval, nzelect, observer, OECD, officedown, officer, ohenery, OmnipathR, opendatatoronto, OptCirClust, optimLanduse, optimParallel, optiSel, OPWeight, ordinalClust, Organism.dplyr, outcomerate, OVESEG, oxcAAR, oxcovid19, packcircles, padma, padr, pafdR, pagenum, pagoda2, paletteer, palettetown, palmerpenguins, pals, pander, parallelDist, parameters, ParamHelpers, paramtest, parcoords, parlitools, parSim, passport, passt, PBImisc, pbixr, pcalg, pcaMethods, pcFactorStan, pct, pder, pdi, PDQutils, pdSpecEst, pdxTrees, PeakError, PeakSegDisk, PeakSegDP, PeakSegJoint, PeakSegOptimal, peptider, perccalc, performance, PerformanceAnalytics, permutes, PheCAP, PhosR, photobiologyFilters, photobiologyInOut, photobiologyLamps, photobiologyLEDs, photobiologyPlants, photobiologySensors, photobiologySun, phyr, physiology, pivottabler, pkgnet, PKNCA, plotscale, plyranges, PNWColors, pocketapi, pointblank, pollen, pollster, pollstR, Polychrome, polymapR, polyRAD, pomp, popEpi, popReconstruct, portalr, povcalnetR, powerlmm, PPforest, pqsfinder, PRDA, prepplot, prevtoinc, proActiv, probably, pROC, profr, profvis, ProjectTemplate, projmgr, PRROC, PSCBS, psd, pseval, pspline.inference, psychonetrics, psymonitor, pwr, pysd2r, qcmetrics, QFeatures, QSutils, qualmap, qualvar, quantdr, QuantNorm, questionr, queuecomputer, R.SamBada, R2admb, r2rtf, r3PG, R3port, r5r, R6, R6P, radarBoxplot, ragtop, raincpc, RandomForestsGLS, rangeMapper, Raquifer, rasterVis, raw, rayshader, rbefdata, rbi, RcmdrPlugin.MA, rco, RcppHungarian, RCzechia, Rdca, rdefra, rdhs, rdi, RDML, rdtLite, Rdtq, readabs, readrba, readsdr, rearrr, recipes, Recon, recountmethylation, REddyProc, regionReport, rem, remap, reporter, representr, retrodesign, reval, revdbayes, revulyticsR, rex, rfordummies, RforProteomics, RGCCA, RGENERATEPREC, rhdf5, rhosa, ricu, Rilostat, riskParityPortfolio, rivr, rKolada, rlc, rLFT, Rmbal, rmcorr, rmdfiltr, rmdformats, rmdpartials, rnbp, rnpn, RoBMA, robustbase, rODE, roloc, rpact, rpanel, rpdo, rpf, rPref, Rpvt, RQEntangle, rrd, Rrelperm, RRphylo, rrscale, rsample, rsetse, rsinaica, rsppfp, rstpm2, rsvd, rsvg, rsyncrosim, rtematres, rtf, Rthingsboard, rtide, RTransferEntropy, rTRM, rtweet, rubias, runstats, rvinecopulib, RWmisc, sadists, safetyGraphics, SAGMM, salem, salesforcer, SAMBA, SamplerCompare, SAMtool, sand, sarks, sbo, SC2API, scales, SCATE, scdhlm, scico, scmeth, scone, scoringTools, scp, scPCA, scry, SDaA, SDEFSR, sdmpredictors, sdmvspecies, seer, seewave, selectKSigs, seqbias, seqhandbook, SEtools, sf, sfcr, sfnetworks, sftrack, sgmcmc, shades, shapr, shiny, shinybrms, shinydlplot, shinyjqui, shinyMobile, shinyPredict, shinyWidgets, SightabilityModel, sigmoid, signnet, signs, SimDesign, simglm, simhelpers, simITS, simplifyEnrichment, simstandard, simstudy, SimSurvey, simTool, simulator, Single.mTEC.Transcriptomes, SingleCaseES, SingleR, sitmo, SiZer, sjmisc, sjstats, slim, slinky, smacof, smcfcs, SMFilter, snifter, socialmixr, sociome, socviz, Sofi, soilDB, SomaticCancerAlterations, sotkanet, SourceSet, sparklyr, sparseMVN, spatialsample, spbabel, spdep, SPONGE, spongebob, spread, srvyr, SSOSVM, ssPATHS, stabm, starm, stars, states, stats19, statsExpressions, stlcsb, StratifiedSampling, streamDepletr, strider, strm, struct, StructFDR, StructuralVariantAnnotation, sugarbag, SunsVoc, SuperLearner, swag, swfdr, swmmr, T4transport, tableone, taipan, TAPseq, taRifx, tbl2xts, tboot, tbrf, tcensReg, TeachingDemos, tealeaves, testassay, testDriveR, texreg, textplot, TFEA.ChIP, tfestimators, TFutils, themis, threesixtygiving, tibble, tictactoe, tidyBF, tidybulk, tidycat, tidycensus, tidyfst, tidygeocoder, tidyhydat, tidyjson, tidylo, tidync, tidytext, tidytransit, tidyxl, tigris, tikzDevice, TimeProjection, tint, tlsh, tmap, Tmisc, tongfen, totalcensus, tourr, tracerer, trackdf, TrafficBDE, TrajDataMining, tramME, transx, trapezoid, treasuryTR, treecm, treedater, treeio, treespace, treestructure, TreeSummarizedExperiment, trending, tRNAscanImport, truelies, ts.extend, tsbox, tsfeatures, tsibble, tsibbledata, tsibbletalk, TSRchitect, TSrepr, tufterhandout, tvgeom, UBL, ukgasapi, ukpolice, unvotes, usdata, UsingR, usmap, uSORT, utilities, vacuum, vagalumeR, VancouvR, VariantAnnotation, vcdExtra, vcfR, vdar, vdiffr, velociraptor, venn, vimp, vinereg, vioplot, vipor, viridisLite, visreg, vlad, volleystat, vote, vroom, vtree, warbleR, WaveSampling, waydown, wbstats, wdpar, weathercan, WeightedROC, wesanderson, whereami, widyr, wildlifeDI, WoodSimulatR, wordbankr, WordR, WRS2, wrswoR, wrswoR.benchmark, xaringanthemer, xgb2sql, xgboost, XICOR, XLConnect, yardstick, zenplots, zinbwave, zoo, zoolog, ztpln, ZVCV
#> 
#> Reverse enhances: flowr, LOLA, mcaGUI, surveillance, tis
#> 
#> 
#> Tip: Use go("ggplot2","manual") to view the manual of package 'ggplot2' and go("ggplot2","website") to visit its website (if any is provided).