Package 'packagefinder'

Comfortable search for R packages on CRAN directly from the R console

What is packagefinder? How does it benefit me?

Currently, there are more than 16,000 R package contributions on CRAN providing R with an unparalleled wealth of features. The downside of the large and increasing amount of packages is that it becomes increasingly difficult to find the right tools to tackle a specific problem. Unfortunately, CRAN does not provide any good search functionality.

packagefinder is designed to search for CRAN packages right from the R console. The philosophy behind this package is that R users like using the R console and need a tool to do their day-to-day-work on CRAN without leaving their normal workspace, the console. In fact, the idea is that with packagefinder you do not need to leave the R console to work with CRAN effectively.

packagefinder is developed to save you time and energy finding the right packages to work with, thereby making your work with R more productive.

Where can I get more information?

Contact the author:

E-mail: Joachim Zuckarelli (joachim@zuckarelli.de)

Twitter: @jsugarelli

packagefinder introduction: A Quick Tutorial

packagefinder on GitHub: https://github.com/jsugarelli/packagefinder

packagefinder video tutorial on YouTube: https://youtu.be/B96NMSo3nJI (discusses version 0.1.5)